Refactor classes and constants
