+++ src/lzdb/constants.py
@@ -1,0 +1,3 @@
+LZDB_REL_DIRECTED = 0
+LZDB_REL_UNDIRECTED = 1
+
